+++ packages/portfolio-contract/test/published-tx-shape.test.ts
@@ -3,40 +3,22 @@ import test from 'ava';
 import { matches } from '@endo/patterns';
 
 import { boardSlottingMarshaller } from '@agoric/client-utils';
-import type { AccountId } from '@agoric/orchestration';
 
-import { TxType, type TxStatus } from '../src/resolver/constants.js';
+import { TxType } from '../src/resolver/constants.js';
 import {
   PublishedTxShape,
-  type PublishedTx,
   type TxId,
+  type PendingTx,
 } from '../src/resolver/types.ts';
+
+import { createMockPendingTxData } from '../tools/mocks.ts';
 
 const marshaller = boardSlottingMarshaller();
 
-type PendingTx = { txId: TxId } & PublishedTx;
-const parsePendingTx = (txId: `tx${number}`, data): PendingTx | null => {
+const parsePendingTx = (txId: TxId, data): PendingTx | null => {
   if (!matches(data, PublishedTxShape)) return null;
   return { txId, ...data } as PendingTx;
 };
-
-export const createMockPendingTxData = ({
-  type = TxType.CCTP_TO_EVM,
-  status = 'pending',
-  amount = 100_000n,
-  destinationAddress = 'eip155:42161:0x742d35Cc6635C0532925a3b8D9dEB1C9e5eb2b64',
-}: {
-  type?: TxType;
-  status?: TxStatus;
-  amount?: bigint;
-  destinationAddress?: AccountId;
-} = {}) =>
-  harden({
-    type,
-    status,
-    amount,
-    destinationAddress,
-  });
 
 test('parsePendingTx creates valid PendingTx from data', t => {
   const txId = 'tx1' as `tx${number}`;
